arch: migrate telegram_bot.services.qdrant + metrics to src.runtime (#2047 slice 1)
Continues the reverse-layering migration tracked under #1948 / #2047 /
#2049. Builds on #2096 (graph.state) and #2099 (graph.config + BGE-M3
embeddings). Moves the QdrantService Smart Gateway and the Prometheus
pipeline-metrics surface into the runtime kernel so src/api/main.py no
longer reaches back into telegram_bot/ for them.

What this PR does:

- Adds src/runtime/services/__init__.py, src/runtime/services/qdrant.py,
  and src/runtime/services/metrics.py with the canonical implementations
  copied verbatim from telegram_bot/services/. The new qdrant module
  imports observability and metrics from src.* (src.observability,
  src.runtime.services.metrics) instead of the telegram_bot.* shims, so
  it satisfies the layering contract.
- Reduces telegram_bot/services/qdrant.py and
  telegram_bot/services/metrics.py to thin re-export shims so every
  existing caller (src/api, telegram_bot internals, plus 25+ test files
  that import from telegram_bot.services.qdrant or .metrics) keeps
  working without churn.
- Switches src/api/main.py:40 to import QdrantService from
  src.runtime.services.qdrant.
- Shrinks tests/data/known_layering_violations.json from 3 entries to 2
  for src/api/main.py: telegram_bot.services.qdrant is now compliant.

Contract test path updates (necessary because canonical home moved):

- tests/contract/test_dead_code_removed_contract.py: 2 _parse(...) calls
  now point at src/runtime/services/qdrant.py for the QdrantService
  AST inspection (mmr_rerank / search_with_score_boosting absence).
- tests/contract/test_error_contract.py: error-span allowlist gains
  src/runtime/services/qdrant.py alongside the legacy shim path.
- tests/contract/test_no_custom_metrics_registry_contract.py: rolling-
  window allowlist gains src/runtime/services/metrics.py; assertion
  message points at src/runtime as the canonical home.
- tests/contract/test_record_counter_metric_migration_contract.py:
  METRICS_OWNER points at src/runtime/services/metrics.py; IMPORT_RE
  matches both new and legacy paths.
- tests/unit/services/test_pipeline_metrics.py: caplog logger updated
  from telegram_bot.services.metrics to src.runtime.services.metrics
  (the logger name follows __name__ at the canonical home).

What this PR does NOT do (out of scope, follow-up work):

- The remaining 2 src/api/main.py imports (telegram_bot.graph.graph,
  telegram_bot.integrations.cache) still flow back into telegram_bot.
  The cache slice is the last large one before #2049 closes; the graph
  slice depends on the cache slice landing first.
- mini_app separability verification is not part of this slice.

Validation:

- uv run --frozen pytest tests/contract/test_layering_no_telegram_bot_imports_contract.py -q
  → 4 passed (allowlist drift detector picks up the shrunk JSON).
- uv run --frozen pytest tests/unit/services/ tests/unit/graph/ tests/contract/ -q
  → 3219 passed (full unit + contract; back-compat shims cover every
  existing caller).
- ruff check + ruff format clean.
- pre-commit gitleaks/end-of-file-fixer/etc. clean.

Refs #2049 #2047 #1948
