+++ tests/data/known_layering_violations.json
@@ -1,8 +1,6 @@
 {
   "src/api/main.py": [
-    "telegram_bot.graph.config",
     "telegram_bot.graph.graph",
-    "telegram_bot.integrations.cache",
-    "telegram_bot.services.qdrant"
+    "telegram_bot.integrations.cache"
   ]
 }
